exp2

+++ scripts/stddev.py
@@ -3,7 +3,7 @@ import numpy as np
 
 nums=[]
 for n in sys.argv[1:]:
-  nums.append(int(n))
+  nums.append(float(n))
 
 a = np.array(nums)
 print str(np.std(a))
